fix bug 118 - file components cannot mirror/rotate
The problem is that the mirrorX and rotated count happens during load
and at that time the number of Ports ins not available.
The number of Ports is only available after recreate, (createSymbol) which
opens subcircuits, verilog and vhdl to count the ports.
Therefore, we skip the Port count for these components.