Corrected simulator process exit status check
Changed slotSimEnded() to check the exit status of the simulator
process, so that the user is informed thru the simulation dialog
when the simulator has crashed.

Replaced some Qt3 compatibility functions with the proper Qt4 ones.
Added newlines at the end of the error messages.
Added some Doxygen style comments.