Make the library search like the Qucs Component search
The Library Tool was using a somewhat cumbersome search, forcing
the user to open a separate dialog for searching.
A search box is now included in the main Library Tool window and
provides an 'instant search' feature, like the main Qucs window
does for the standard component search.
The old Search Dialog widget files are removed.