Add setter for Schematic::FileInfo member variable.
There is a getter function but the FileInfo does not seem to be assigned
anywhere. It is used to keep file information about the file which
stores the schematic. The schematic file that is loaded.

The getFileInfo was returning the default empty string. It is used by
subcircuits, spicefile and  the component dialog to a parent schematic.

Things work fine if the schematic and subcircuit are on the same
directory and QucsHomeDir is set properly. If not, as in the case of
loading a schematic from the command line, it fails to find the parent
schematic and as a result the 'child' subcircuit is not found.

See issue #333