Fix crash when trying to netlist from the CLI.
Running Qucs from the command line to convert an schematic to a qucsator
netlist could lead to a crash.

When running from the command line QucsMain is never in scope, similarly
the QucsMain->schNameHash (related to the search path lookup feature) is
only available on the interative GUI.

This fixes issue #333 for the subcircuit component.