Fix OSX uninstall script.
For some reason the package is recording the installed files with
a leading . in front. Causing the rm command not to find anything.

The file path reported by the BOM file is adjusted so the files can
be found on the correct location.

Also changed the script mode to executable.

[ci skip]