Make "Select default simulator" dialog save changes
A missing "saveApplSettings" call at the end of the slotApply slot means
that changes made to the default simulator setting do not get saved in
the persistent configuration until after other settings dialogs (such as
File->Application Settings...) are opened and their apply slots
triggered.

This missing call has several implications:
  -- The ~/.config/qucs/qucs.conf file is NOT created if a user opens
     qucs-s for the very first time, having never run Qucs before, even
     if one follows the explicit instructions in the "No default
     simulator is set!" dialog.
  -- The only way to get settings saved is by opening File->Application
     Settings... and clicking "Ok"
  -- If a user changes simulator settings later on, the change does not
     get saved, again until the other dialog is opened.

By adding a call to saveApplSettings() at the end of slotApply, the
program behaves as expected --- it saves the settings in persistent
store immediately.