Restore qucslib_common.h
In commit a3842194 qucs-lib was removed from the source tree, and all
references to qucslib_common.h were changed from
"../qucs-lib/qucslib_common.h" to where it appeared it was intended
that qucslib_common.h was supposed to have been moved.

But qucslib_common.h was apparently never committed to the current
branch, and the code on that branch will not build.

This commit simply copies that file off of the master branch and into
current so the code can compile.  I have confirmed that after copying
this file into the qucs directory the current branch builds just fine.

closes #171