feat: Add zooming around point under the mouse pointer
When schematic is zoomed in or out using mouse wheel and Ctrl
button, the drawing is resized around a "focal" point under the
mouse pointer. The "focal" point of the schematic remains
stable relative to the viewport, while other parts of the schematic
get into or out of the viewport according to the zoom level.