fix: Don't ignore newlines in texts when calculating element's width
Newlines in the text of schematic element's property should be treated
as line breaks otherwise the text of an element would become a single
very long line and the calculated width of an element would be incorrect.

This commit fixes such problem by replacing newlines-ignorant method with
a newlines-aware one.

Fixes ra3xdh/qucs_s#426