Remove correction factor from Component::entireBounds
In short: this commit removes coupling between Component
and Schematic in the process of calculating component's
bounds.

  Why it is safe to decouple them

At first, carefully read the comment in Schematic::textCorr.
It describes clearly why the return value of function is equal
to fraction:

                1                              1
  —————————————————————————————————   =   ———————————
  QucsSettings.font.pointSize() * k       lineSpacing

This value is passed to Component::entireBounds every time it's
called.

Inside Component::entireBounds it was used to calculate total
height of component's text properties using this formula:

    textPropertiesCount / Corr

After substituting the value of Corr to formula it yields:
    textPropertiesCount / Corr =
  = textPropertiesCount / (1 / lineSpacing) =
  = textPropertiesCount * lineSpacing
  = textPropertiesCount * QucsSettings.font.pointSize() * k

1. When text properties of a component are one-liners

In this case 'textPropertiesCount * lineSpacing' is equal to the value
of 'totalTextPropertiesHeight', which is calculated in
Component::textSize. We can use 'totalTextPropertiesHeight' directly

2. When text properties of a component are *not* one-liners

In this case the value of 'textPropertiesCount * lineSpacing' is just
wrong because it doesn't account line breaks.

In total this correction seems to be useless. And if 'Corr' value is not
needed in 'Component::entireBounds', then there is no need to couple
invocations of with 'Schematic::textCorr'