Replace usages of sizeOfAll with updateAllBoundingRect
Passing Used.* members to 'sizeOfAll' function is clumsy, it
requires the reader to know the purpose of Used.* variables and
that 'sizeOfAll' function updates its arguments in-place. Then
both of facts must be summed up to understand that the line

 Doc->sizeOfAll(Doc->UsedX1, Doc->UsedY1, Doc->UsedX2, Doc->UsedY2)

means 'update the bounds of the area occupied by elements'

On the other hand, the newly added 'updateAllBoundingRect' clearly
states what it does, it hides access to Used.* members and it's
easier to read and understand.