Extract "Ellips" primitive from a generic "Area"
Both ellipses and rectangles are drawn from a single object named
"Area". Drawing logic looks from where it got the object: if it's
from "rectangles" container, then draw a rectangle, otherwise
draw an ellipse.

This commit creates new "Ellips" object type, separate from "Area".
It opens a way to:
- hide a dedicated ellipsis rectangle drawing logic in a method
  associated only with "Ellips" objects
- distinguish ellipses and rectangles not by their place of residence,
  but by their nature.

The name "Ellips" without "e" at the end is used because "Ellipse"
conflicts with a painting named that way, which is declared in the
same namespace. I didn't want to go into any additional trouble
and chose to go with this type of a "workaround"