Extract "Ellips" primitive from a generic "Area"
Both ellipses and rectangles are drawn from a single object named
"Area". Drawing logic depends on the source of an object: if it's
from "rectangles" container, then draw a rectangle, otherwise
draw an ellipse.

This commit creates new separate "Ellips"  type, which allows to:
- hide a dedicated ellipsis rectangle drawing logic in a method
  associated only with "Ellips" objects
- distinguish ellipses and rectangles by their types, not by the
  place they come from

The name ("Ellips") has no "e" at the end because "Ellipse"
conflicts with a painting with the same name, which is declared in
the same namespace. I didn't want to go into any additional trouble
and chose to go with this type of a "workaround".