Allow inserting ports when editing a symbol
"Insert port" tool is reused it. When in symbol editing mode,
left mouse button click with this tool activated spawns a dialog
asking for port name, then insert a port.

I hope there is no breaking changes and everything is backwards
compatible as I strived to add workarounds everywhere it's needed.

The sad part is that code is not elegant, it increases overall
codebase entropy, doesn't follow any general strategy. It's just
a set of patches.