Fix tabdiagram headers
It was found that line #250 filters the dependent variable name by removing all characters before the "/". This makes sense if the data comes from an NGspice simulation. However, when the "Qucsator" simulator is selected, the variable names come without the "/".

This commit fixes this problem by first checking if the string contains "/"