Beautify the attenuator schematics
This commit is intented to produce more human-readable schematics

Before this commit the attenuator schematics were too compact and it was difficult (or even impossible) to read the resistor values at first sight.

Also, the resistor values were rounded to 1 decimal place as it is more than enough for actual attenuator design.