Added SPAR simulation checkbox
It was added a checkbox in the qucsatenuattor UI in order to let the user to attach a SP simulation component to the attenuator circuit.

Resistive attenuators are wideband circuits so the GUI doesn't ask for a frequency sweep. 

The tool sets the SP sweep to 50MHz-3GHz. If needed, the user can change that before running the simulation.