Fix detached ports disappearing after undo in symbol edit mode
Detached ports are ports which were inserted manually, i.e. they
don't have a corresponding port in schematic.

Inside Schematic::undo the Schematic::adjustPortNumbers is called.
The latter removes detached ports and because of that you lose
all of your ports if use "undo" once.

But Schematic::adjustPortNumbers is also called every time the
editing mode is switched and when the file is saved. So I think
it wouldn't be harm to not adjust port numbers every time you
undoing.