Unify schematic generation code for MS and ideal TLIN implementation
The schematic generation code for microstrip and ideal transmission lines splitted at some point, duplicating the for-loop, etc. The code is much clearer and less prone to errors if both the microstrip implementation and the TLINES share the same logic as much as possible.

It was also found that in the microstrip implementation, the lambda/4 distance did not take into account the effective dielectric permittivity.

The recently created functions getZ() and getMicrostripWidth() were removed. Their functionality is implemented in a single line in the createSchematic() reducing the clutter.