Move schematic drawing functions to Filter class
getLineString(), getTeeString(), and getWireString() were recently created inside the QuarterWave_Filter class. The purpose of these functions is to ease the schematic generation code and they could be used in other filter implementations. It makes more sense to put them in a more general class, such as the Filter class.