Restore qpainter state after paining the id_text element
The 'paint' function of ID_text missed the restore() call on painter,
which led to carrying the changes made to painter further down the
call chain. In other words, the painter state *after* painting the
ID_text wasn't the same as *before* doing it, and everything painted
after the ID_text was affected by this change in painter's state.

This commit add missing restore() call.

Fixes: ra3xdh/qucs_s#739