Ensure Zout = Zin (=Z0) in Lpad and QW attenuators
It was found that when selecting the QW or L-pad attenuators and marking the option to include the S-parameter analysis block, the impedance of the output term was not Z0 if Z0 != 50

This happened because these attenuator topologies were missing from slotSetText_Zin( double val ). This function calls the calculation function in order to update the power dissipation data whenever the user changes Z0.