Fix issue #905
As reported in issue #905, the matching network tool gave wrong results when using a conditionally stable S-matrix. In order to have a conjugately match, the device is required to meet the unconditional stability criteria defined in [1] as K > 1 and abs(delta) < 1

A condition was added in the program to filter out these cases and warn the user against this limitation.

[1] Microwave Engineering. David M Pozar. 2012. Eqs 12.28 and 12.29