Fix: Widgets overlap when after removing trace/limit/marker
It was found that after a trace/limit/marker is removed, if the user adds a new trace/limit/marker, the new widgets overlap the last ones. 

This happened because the position of the widgets in the QGridLayoyt were not updated after a trace/limit/marker is removed.

This commit addresses this issue.