Improvement: Save only S-par data on session files
Before this commit, all traces associated to a S-parameter file were saved in the session file even if the user was not displaying them. This causes a heavy overhead specially in large files.

Now, the session file only contains S-parameter data. If the user wants to display additional traces such as K, mu, |delta|, etc., they are computed on demand