Fix netlist generation after placing the component instances from newOne() placement
It was found that the netlist generated from components placed using the Component::newOne() function was wrong. This was caused by the copy constructor.

This type of placement ocurrs when the user picks a component from the palette, place the component in the schematic and Qucs-S automatically puts another one, grayed, waiting to be placed.

Also, it was found some minor error with the XML component definition and a problem with the placement of the GND.

At this point, Qucs-S works well for the example of a voltage divider (at least with Qucsator)