Slight change in Qucs-S schematic file format
It is possible to have different components with the same name in different libraries. For example, "R" can be the ideal resistor from the Lumped Components library, but it can also be a custom resistor from a PDK. To avoid this kind of ambiguity, the netline of the component now includes the library according to the following format Library:ComponentName. Both "Library" and "ComponentName" must not contain spaces, so they are replaced by "_".

Backward compatibility with old Qucs-S schematics is ensured.