Add setters for coordinates of ends of a wire
While it's possible to change coordinates of wire ends (and therefore
dimensions of the wire) by directly assigning to its x1, y1, etc.
members, it has a disadvantage of not accounting the presence of
a WireLabel on the wire. Such changes may lead to label root "hanging
in the air" and not to pointing to the wire.

Setters address this problem and keep everything in tact.