Fix crash due to geometry invariant violation
In some cases dangling wires could appear because of how dumb
connect with worked. While creating a new node it could split a
wire marked for deletion which led to deleting only one segment
of such wire. With this commit nodes are created in "dumb" way,
without any checks and split.