Stop forcing LEVEL=2 into Xyce diode conversions
Qucs diode models use some parameters that the Xyce level 1 diode did
not support prior to 2022, and the "PSpice-compatible" level 2 diode
did have those parameters.

Xyce's level 1 diode has been reworked to include those parameters and
more, and most of the differences between level 1 and level 2 are
gone.  In fact, the only differences that remain are in the
computation of junction voltage and junction capacitance, and it is
believed that they are incorrect anyway.

This commit removes the code that forces LEVEL=2 into diode models
converted from Qucs models to Xyce netlists.

Closes GH-1372