Limit the multiple-file opening feature to files
The top-level items are entered using the APPEND_ROW macro.
In this commit, APPEND_ROW macro was modified to remove the Qt::ItemIsSelectable flag.

This prevents top-level items (e.g. "Schematic", "Others" group labels) to be selectable. Since they have two columns, the flag need to be removed in the title (column 0) as well as in the "Notes" section (column 1), which is always blank