Remove label type constants in favour of single 'isLabel'
- remove macros isHWireLabel, isHMovingLabel, etc.
- always use 'isLabel' type for labels

Some of the removed macros, such as isHMovingLabel are pre-#1266
legacy and are no longer in use.

The distinction between HWireLabel and VWireLabel is relevant only
during painting a label and can be inferred directly from the host
wire of the label. No need to keep dedicated label types for that.

Node and wire labels can be distinguished via their respective host
elements, accessible through "owner()".