S-parameter Viewer: Fix trace count in the magnitude / phase display
When adding phase plots, it was found that the control widgets overlapped with those of the magnitude. This occurred because the trace count was incorrect. Since this chart is used for both magnitude and phase traces, the number of traces must be counted before adding a trace to the chart.