Connect ComponentDialog::slotOKButton() to QDialog::accepted().
Instead of connecting directly the QDialogButtonBox::accepted() signal
to ComponentDialog::slotOKButton(), connect it to QDialog::accept()
and then connect QDialog::accepted() to the slotOKButton() method.

This effectively splits the accepted signal delivery in two steps:
from QDialogButtonBox to QDialog and from QDialog to slotOKButton().

On some systems, this allows the cell editors of the property table to
store any pending changes before reading back the data and returning
to the Qucs-S editor.

Fixes #1453.