S-par Viewer [Matching]: Adjust traces display
The traces were clearer only when the user switches the tool tab. However, the Matching tab can toggle two modes: 1-port and 2-port matching. These modes are set to display different traces and keeping the 1-port matching trace when the 2-port matching is activated clutters a lot the display.

A dedicated function was created to remove the datasets and clear their associated traces. An optional argument is added so that a dataset can be excluded