S-par Viewer [Filtering]: Add relative bandwidth checks for some BPF topologies
Some BPF implementations are intrinsically narrowband. The relative BW must be checked for ensuring a successful implementation. In this sense, whenever the user selects a new topology a function is called to adjust the bandwidth if the maximum relative bandwidth setting is exceeded.

This check is run only when the topology is selected for the first time, allowing the user to try wider rel BW later.