S-par Viewer [Attenuator]: Refactor attenuator designer
The AttenuatorDesigner class was just a wrapper for accessing the design code. This is actually not needed. Following the same approach as done with the Filter Tool, the Attenuator tool is now modular so that each attenuator topology is contained in a class. This forced the redesign of how the synthesis process works