Update open file icons
The text inside the open file icons is tiny and it's difficult to distinguish between the new symbol file and the new text file.

The texts were put in black and in capital letters.
The shadow layers were removed

In case of the new schematic file, the symbol of a NPN transistor [1] was added.

[1] NPN symbol:
https://commons.wikimedia.org/wiki/File:Electrical_symbols_library.svg

Author: https://commons.wikimedia.org/wiki/User:FDominec