S-par Viewer [DC Filters]: Fixes and simplifications
Before this commit, it was possible to specify either the inductance or the capacitance of the resonators as design parameter. However, the design procedure modifies the L or the C of the resonator to absorb the inverter capacitance/inductance, depending on the topologies. This may lead to confusion to the user as it will see that the synthesized filter may not show the inductance/capacitance it entered.

For example, in the capacitive-coupled shunt resonator topology, the user enter the resonator inductances as design parameter and then the coupling and the capacitive inverters are calculated (see Matthaei's book). This leads to a modify the shunt capacitance to absorb the inverter's negative C. 

If the user specified a capacitance, it would actually see that after the synthesis the resonators have a different capacitance as it specified. I decided to keep it simple as in the reference.