fix(ngspice/noise): save nodes relative to plot number
This change uses the relative plot to save the correct
nodes. We use 'setplot noiseX' and then use the relative
wildcard save 'all' to get the equivalent of 'noiseX.all'.

Previosuly, when using the fixed value of
'noise1.all' it would end up using the first result for every plot.
Making it seems like there was one singular plot in the case of sweeps.

Fixes: 09923a89 ("plot noise contributors")