Fix issue #1551: Duplicate and Rename actions disabled in Project view
The "Duplicate" and "Rename" actions were disabled when the user selected a single schematic file in the project view panel.

The project widget contains two columns: the file name and a note. When the user selected a file, the first and the second columns are automatically selected, so selectionModel->selectedIndexes().count() = 2. Consequently, "multipleSelected" variable was true, and the "Duplicate" and "Rename" actions were hidden.

The problem is fixed by using "selectedRows()" instead of "selectedIndexes()"