Diagram Dialog: Show trace properties in the list of graphs
Before this commit, the graph list contained only the name of the traces and the user needed to click on each trace to see the trace properties.

In this commit, the graph list is transformed into a table. This allows to show the properties of each trace in a single row so the user can inspect the properties of all of them at a glance