Diagram Dialog: Refactor to remove Property2 QLineEdit
Property2 QLineEdit was used for setting the number of decimal precision digits in tabular data and thickness in trace graphs at the same time. in order to improve readability and maintainability, this functionality was split into two separate widgets: one QSpinBox for setting the trace thickness and another QSpinBox for indicating the number of decimal positions.

The visibility of these widgets is adjusted depending on the type of graph